+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Version 0.2.2
 
 * Fixed dragonruby install from itch.io package
+* Fixed issue with add with empty dependencies in config
+* Fixed `smaug.rb` generation on Windows